refactor: 코드리뷰 일부반영
