+++ 01_todo/suKyoung/src/types.ts
@@ -1,0 +1,9 @@
+export type HTMLElements = {
+  form: HTMLFormElement;
+  inputSearch: HTMLInputElement;
+  buttonSubmit: HTMLInputElement;
+  buttonEdit: HTMLInputElement;
+  buttonDeleteAll: HTMLInputElement;
+  listResult: HTMLUListElement;
+  messageAlert: HTMLParagraphElement;
+}
